refactoring routers
