+++ app/main.py
@@ -10,6 +10,7 @@ from sqlalchemy.orm import Session
 from . import models, schemas, utils
 from .database import engine, get_db
 import sqlalchemy.dialects.postgresql 
+from .routers import user, post
 
 
 
@@ -41,77 +42,14 @@ def find_index_post(id):
     for i, p in enumerate(my_post):
         if p['id'] == id:
             return i  
-          
+ 
+
+app.include_router(post.router)
+app.include_router(user.router)
+         
 @app.get("/")
 def root():
     return {"message":"Hello World"}
 
 
-@app.get("/posts", response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
-    # cursor.execute(""" SELECT * FROM posts""")
-    # posts = cursor.fetchall()
-    posts = db.query(models.Post).all()
-    return  posts
-
-@app.post("/posts", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db)):
-    # cursor.execute("""INSERT INTO posts (title, content, published) VALUES (%s, %s, %s) RETURNING * """, (post.title, post.content, post.published))
-    # new_post = cursor.fetchone()
-    # conn.commit()
-    new_post = models.Post(**post.dict())
-    db.add(new_post)
-    db.commit()
-    db.refresh(new_post)
-    return new_post
-
-@app.get("/posts/{id}", response_model=schemas.Post)
-def get_post(id: int, db: Session = Depends(get_db)):
-    # cursor.execute(""" SELECT * FROM posts WHERE id = %s""", (str(id)))
-    # post = cursor.fetchone()
-    post = db.query(models.Post).filter(models.Post.id == id).first()
-    if not post:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
-                            detail=f"post with id: {id} was not found")
-    return post
-
-
-@app.delete("/posts/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id:int, db: Session = Depends(get_db)):
-    # cursor.execute("""DELETE FROM posts WHERE id = %s RETURNING *""", (str(id)))
-    # delete_post = cursor.fetchone()
-    # conn.commit()
-    post = db.query(models.Post).filter(models.Post.id == id)
-    if post.first() == None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"post with id {id} does not exist")
-    post.delete(synchronize_session=False)
-    db.commit()
-    
-@app.put("/posts/{id}", response_model=schemas.Post)
-def update_post(id:int, updated_post: schemas.PostCreate, db: Session = Depends(get_db)):
-    # cursor.execute("""UPDATE posts SET title = %s, content = %s, published = %s  WHERE id = %s RETURNING *""", (post.title, post.content, post.published, str(id)))
-    # updated_post = cursor.fetchone()
-    # conn.commit()
-    post_query = db.query(models.Post).filter(models.Post.id == id)
-    post = post_query.first()
-    if post == None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"post with id {id} does not exist")
-        
-    post_query.update(updated_post.dict(),synchronize_session=False)
-    db.commit()
-    return  post_query.first()
-
-@app.post("/users", status_code=status.HTTP_201_CREATED, response_model=schemas.UserOut)
-def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
-    #hash the password - user.password
-    hashed_password = utils.hash(user.password)
-    user.password = hashed_password
-    new_user = models.User(**user.dict())
-    db.add(new_user)
-    db.commit()
-    db.refresh(new_user)
-    return new_user
-
